backend/requirements.txt
django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework-simplejwt==5.4.0
google-genai>=0.8.0
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
